requirements.txt
google-generativeai>=0.1.0
openai
youtube-transcript-api
Flask
pyyaml
Flask-Injector
tqdm